moved the annotation check routines into the generator
