+++ javaslang-match/src/main/java/javaslang/match/annotation/Patterns.java
@@ -5,8 +5,6 @@
  */
 package javaslang.match.annotation;
 
-import javax.annotation.processing.Messager;
-import javax.lang.model.element.TypeElement;
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
@@ -21,23 +19,4 @@ import java.lang.annotation.Target;
 @Target(ElementType.TYPE)
 @Retention(RetentionPolicy.CLASS)
 public @interface Patterns {
-
-    /**
-     * Checks if an {@link javax.lang.model.element.TypeElement} is a valid {@code @Patterns} type.
-     */
-    class Checker extends BaseChecker<Checker, TypeElement> {
-
-        public static boolean isValid(TypeElement typeElement, Messager messager) {
-            final Checker typeChecker = new Checker(typeElement, messager);
-            return typeChecker.ensure(e -> !e.isNested(), () -> "Patterns need to be defined in top-level classes.");
-        }
-
-        private Checker(TypeElement elem, Messager messager) {
-            super(elem, messager);
-        }
-
-        boolean isNested() {
-            return elem.getNestingKind().isNested();
-        }
-    }
 }
