+++ javaslang-match/src/main/java/javaslang/match/generator/BaseChecker.java
@@ -3,7 +3,7 @@
  *  _/  /  /\  \  \/  /  /\  \\__\\  \  //  /\  \ /\\/ \ /__\ \   Copyright 2014-2016 Javaslang, http://javaslang.io
  * /___/\_/  \_/\____/\_/  \_/\__\/__/\__\_/  \_//  \__/\_____/   Licensed under the Apache License, Version 2.0
  */
-package javaslang.match.annotation;
+package javaslang.match.generator;
 
 import javax.annotation.processing.Messager;
 import javax.lang.model.element.Element;
